유니티의 HTTP 요청 처리하기 : 게임 캐릭터 생성하기 완료
GameController : 게임 캐릭터 생성을 위한 메서드를 추가하였다.
GameCharacterService : 게임 캐릭터 생성 서비스를 추가하였다. (생성을 한후 서버로부터 다시 가져오는 방식으로 잘 저장 됬는지 확인을 위해 사용된다.)

GameCharacterViewController, GameCharacter, GameHighScore, GetGameCharacterDTO, UpdateGameCharacterHighScoreDTO
게임 캐릭터와 랭크 도메인의 최고 점수의 변수형이 int 로 잘못되 있어 Long 형으로 전체 편집하였다.

+++ gamewebpage/src/main/java/Park/gamewebpage/domain/GameCharacter.java
@@ -25,10 +25,10 @@ public class GameCharacter {
     
     // 향후 필요 없어 보이면 없엘것
     @Column(name = "high_score")
-    private int highScore;
+    private Long highScore;
 
     @Builder
-    public GameCharacter(String email, String nickname, int highScore){
+    public GameCharacter(String email, String nickname, Long highScore){
         this.email = email;
         this.nickname = nickname;
         this.highScore = highScore;
